Move debug functions to own module
This commit is part of efforts to isolate subsystems of project into own modules.
