+++ Descent3/CMakeLists.txt
@@ -297,7 +297,7 @@ file(GLOB_RECURSE INCS "../lib/*.h")
 
 add_executable(Descent3 ${HEADERS} ${CPPS} ${PLATFORM_CPPS} ${INCS})
 target_link_libraries(Descent3
-  2dlib AudioEncode bitmap cfile czip d3music ddio_common libmve libacm
+  2dlib AudioEncode bitmap cfile czip d3music ddebug ddio_common libmve libacm
   fix grtext manage mem misc model module movie stream_audio
   music networking physics renderer rtperformance sndlib ui unzip vecmat md5
   ${PLATFORM_LIBS})
