+++ Descent3/rocknride.cpp
@@ -42,7 +42,7 @@
  */
 #include "mono.h"
 #include "ddio.h"
-#include "debug.h"
+#include "ddebug/debug.h"
 #include "game.h"
 #include "pserror.h"
 #include "rocknride.h"
