+++ ddebug/macdebug.cpp
@@ -56,7 +56,7 @@
  * $NoKeywords: $
  */
 
-#include "debug.h"
+#include "ddebug/debug.h"
 #include "mono.h"
 #include <assert.h>
 #include <stdarg.h>
