+++ lib/pserror.h
@@ -148,7 +148,7 @@
 #ifndef PSERROR_H
 #define PSERROR_H
 #include <assert.h>
-#include "debug.h"
+#include "debug/debug.h"
 #include "mono.h"
 //	initializes error handler.
 bool error_Init(bool debugger, bool mono_debug, const char *app_title);
